stop using sudo for Windows in CI

+++ test/storage/remove-empty-state.sh
@@ -10,7 +10,9 @@ test ! -s "$LOCAL_STATE_DIR/status" # must be empty
 rm "$LOCAL_STATE_DIR/status"
 rmdir "$LOCAL_STATE_DIR/"
 
-sudo test -e "$GLOBAL_STATE_DIR/status"
-sudo test ! -s "$GLOBAL_STATE_DIR/status"
-sudo rm "$GLOBAL_STATE_DIR/status"
-sudo rmdir "$GLOBAL_STATE_DIR/"
+if [ "$USE_SUDO" = true ]; then
+    sudo test -e "$GLOBAL_STATE_DIR/status"
+    sudo test ! -s "$GLOBAL_STATE_DIR/status"
+    sudo rm "$GLOBAL_STATE_DIR/status"
+    sudo rmdir "$GLOBAL_STATE_DIR/"
+fi
